Implemeted delete product by id

+++ factory/src/main/java/com/josephadogeridev/factory/product/ProductService.java
@@ -4,13 +4,18 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestBody;
+
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @Service
 public class ProductService {
 
     private final ProductRepository productRepository;
+    Map<String, String> responseBody ;
+
 
     public ProductService(ProductRepository productRepository) {
         this.productRepository = productRepository;
@@ -31,18 +36,27 @@ public class ProductService {
 
 
     public Product findProductById(Long id) {
+
         return productRepository.findById(id).orElse(null);
     }
 
-    public void deleteProduct(Long id) {
+    public ResponseEntity<?> deleteProduct(Long id) {
+        System.out.println("product delete id: " + id);
+        Product product = productRepository.findById(id).orElse(null);
+
+        if (product == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         productRepository.deleteById(id);
+        responseBody = new HashMap<>();
+        responseBody.put("message", "Successfully deleted product with id: " + id);
+        return new ResponseEntity<>(responseBody, HttpStatus.BAD_REQUEST);
     }
 
     public ResponseEntity<?> createProduct(Product product){
-        System.out.println("inside create product");
         Optional<Product> productOptional = productRepository.findByName(product.getName());
         if (productOptional.isPresent()) {
-            throw new IllegalStateException("Email already exists");
+            throw new IllegalStateException("product name already exists");
         }
         productRepository.save(product);
